requirements.txt
openai
google-cloud-texttospeech
sounddevice
scipy
SpeechRecognition
python-dotenv
pyaudio